scraper/requirements.txt
celery
redis
celery
cassandra-driver
fastapi
python-dotenv
fake-useragent
selenium
requests_html
notebook